[SW-33] move: 파일 위치 수정
