+++ data/src/main/java/com/example/data/repository/MainRepositoryImpl.kt
@@ -2,8 +2,8 @@ package com.example.data.repository
 
 import com.example.data.service.SubwayService
 import com.example.domain.repository.MainRepository
-import com.example.domain.repository.model.LineInfoResponse
-import com.example.domain.repository.model.StationDetailResponse
+import com.example.domain.model.LineInfoResponse
+import com.example.domain.model.StationDetailResponse
 import javax.inject.Inject
 
 class MainRepositoryImpl @Inject constructor(
