+++ domain/src/main/java/com/example/domain/repository/MainRepository.kt
@@ -1,7 +1,7 @@
 package com.example.domain.repository
 
-import com.example.domain.repository.model.LineInfoResponse
-import com.example.domain.repository.model.StationDetailResponse
+import com.example.domain.model.LineInfoResponse
+import com.example.domain.model.StationDetailResponse
 
 interface MainRepository {
     suspend fun getAllLineInfo(): Result<LineInfoResponse>
